// packages/monaco-lsp-client/src/types.ts
export type TextDocumentContentChangeEvent =
  | { range: Range; rangeLength?: number; text: string }
  | { text: string };
// ...
export interface Range {
  /**
   * The range's start position.
   */
  start: Position;
  /**
   * The range's end position.
   */
  end: Position;
}
// ...
export interface TextDocumentIdentifier {
  /**
   * The text document's uri.
   */
  uri: string;
}
// ...
export interface Position {
  /**
   * Line position in a document (zero-based).
   *
   * If a line number is greater than the number of lines in a document, it defaults back to the number of lines in the document.
   * If a line number is negative, it defaults to 0.
   */
  line: number;
  /**
   * Character offset on a line in a document (zero-based).
   *
   * The meaning of this offset is determined by the negotiated
   * `PositionEncodingKind`.
   *
   * If the character value is greater than the line length it defaults back to the
   * line length.
   */
  character: number;
}